fix(rename): integrated WriteLock on rename coomand
