+++ zboxcore/sdk/allocation.go
@@ -868,6 +868,7 @@ func (a *Allocation) RenameObject(path string, destName string) error {
 	}
 
 	req := &RenameRequest{}
+	req.allocationObj = a
 	req.blobbers = a.Blobbers
 	req.allocationID = a.ID
 	req.allocationTx = a.Tx
